requirements.dev.txt
mypy==1.8.0
mypy-extensions==1.0.0
pylint==4.0.4
pylint-django==2.7.0
pytest==9.0.2
pylint-django==2.7.0
types-Pillow==10.2.0.20240111
types-pytz==2025.2.0.20250516
types-PyYAML==6.0.12.12
typing_extensions==4.13.2
